+++ yt_dlp/extractor/extractors.py
@@ -1,8 +1,8 @@
 import inspect
 import os
 
-from .._globals import LAZY_EXTRACTORS
-from .._globals import extractors as _extractors_context
+from ..globals import LAZY_EXTRACTORS
+from ..globals import extractors as _extractors_context
 
 _CLASS_LOOKUP = None
 if not os.environ.get('YTDLP_NO_LAZY_EXTRACTORS'):
